Split two packages
