+++ cameraview/src/main/java/com/otaliastudios/cameraview/filters/FillLightFilter.java
@@ -2,6 +2,8 @@ package com.otaliastudios.cameraview.filters;
 
 import androidx.annotation.NonNull;
 
+import com.otaliastudios.cameraview.filter.BaseFilter;
+
 /**
  * Applies back-light filling to the preview.
  */
