+++ cameraview/src/main/java/com/otaliastudios/cameraview/filters/InvertColorsFilter.java
@@ -2,6 +2,8 @@ package com.otaliastudios.cameraview.filters;
 
 import androidx.annotation.NonNull;
 
+import com.otaliastudios.cameraview.filter.BaseFilter;
+
 /**
  * Inverts the preview colors. This can also be known as negative Effect.
  */
