+++ cameraview/src/main/java/com/otaliastudios/cameraview/filters/SaturationFilter.java
@@ -2,6 +2,8 @@ package com.otaliastudios.cameraview.filters;
 
 import androidx.annotation.NonNull;
 
+import com.otaliastudios.cameraview.filter.BaseFilter;
+
 /**
  * Adjusts color saturation of preview.
  */
