+++ cameraview/src/main/java/com/otaliastudios/cameraview/internal/egl/EglViewport.java
@@ -7,7 +7,7 @@ import android.opengl.GLES20;
 import androidx.annotation.NonNull;
 
 import com.otaliastudios.cameraview.CameraLogger;
-import com.otaliastudios.cameraview.filters.Filter;
+import com.otaliastudios.cameraview.filter.Filter;
 import com.otaliastudios.cameraview.filters.NoFilter;
 import com.otaliastudios.cameraview.internal.GlUtils;
