+++ cameraview/src/main/java/com/otaliastudios/cameraview/preview/FilterCameraPreview.java
@@ -7,7 +7,7 @@ import android.view.ViewGroup;
 
 import androidx.annotation.NonNull;
 
-import com.otaliastudios.cameraview.filters.Filter;
+import com.otaliastudios.cameraview.filter.Filter;
 
 
 /**
