+++ demo/src/main/java/com/otaliastudios/cameraview/demo/CameraActivity.java
@@ -27,7 +27,7 @@ import com.otaliastudios.cameraview.PictureResult;
 import com.otaliastudios.cameraview.controls.Mode;
 import com.otaliastudios.cameraview.VideoResult;
 import com.otaliastudios.cameraview.controls.Preview;
-import com.otaliastudios.cameraview.filters.Filters;
+import com.otaliastudios.cameraview.filter.Filters;
 import com.otaliastudios.cameraview.frame.Frame;
 import com.otaliastudios.cameraview.frame.FrameProcessor;
 
@@ -35,8 +35,6 @@ import java.io.ByteArrayOutputStream;
 import java.io.File;
 import java.util.Arrays;
 import java.util.List;
-
-import static com.otaliastudios.cameraview.filters.Filters.*;
 
 
 public class CameraActivity extends AppCompatActivity implements View.OnClickListener, OptionView.Callback {
